% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssx-package.R
\docType{package}
\name{gamlssx-package}
\alias{gamlssx}
\alias{gamlssx-package}
\title{gamlssx: Generalised Additive Extreme Value Models for Location, Scale and
Shape}
\description{
Describe
}
\details{
More details.
}
\references{
Rigby R.A. and Stasinopoulos D.M. (2005). Generalized additive
models for location, scale and shape (with discussion), \emph{Appl. Statist.},
\strong{54}, part 3, pages 507-554. \doi{10.1111/j.1467-9876.2005.00510.x}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulnorthrop.github.io/gamlssx/}
  \item \url{https://github.com/paulnorthrop/gamlssx}
  \item Report bugs at \url{https://github.com/paulnorthrop/gamlssx/issues}
}

}
\author{
\strong{Maintainer}: Paul J. Northrop \email{p.northrop@ucl.ac.uk} [copyright holder]

Authors:
\itemize{
  \item Jennifer Ji [copyright holder]
}

}
