% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGEV.R
\name{fitGEV}
\alias{fitGEV}
\title{Fit a Generalized extreme value (GEV) GAMLSS}
\usage{
fitGEV(formula, stepLength = 1, stepAttempts = 2, stepReduce = 2, ...)
}
\arguments{
\item{formula}{Argument \code{formula} passed to \code{\link[gamlss:gamlss]{gamlss}}.}

\item{stepLength}{A numeric vector containing positive values. The initial
values of the step lengths \code{mu.step}, \code{sigma.step} and \code{nu.step} passed to
\code{\link[gamlss:gamlss.control]{gamlss.control}} in the first attempt to fit
the model by calling \code{\link[gamlss:gamlss]{gamlss}}. If \code{stepLength} has a
length that is less than 3 then \code{stepLength} is recycled to have length 3.}

\item{stepAttempts}{A non-negative integer. If the first call to
\code{\link[gamlss:gamlss]{gamlss}} throws an error then we make \code{stepAttempts}
further attempts to fit the model, each time dividing by 2 the values
of \code{mu.step}, \code{sigma.step} and \code{nu.step} supplied to
\code{\link[gamlss:gamlss.control]{gamlss.control}}. If \code{stepAttempts < 1} then
no further attempts are made.}

\item{stepReduce}{A number greater than 1. The factor by which the step
lengths in \code{stepLength} are reduced for each extra attempt to fit the
model. The default, \code{stepReduce = 2} means that the step lengths are
halved for each extra attempt.}

\item{...}{Further arguments passed to \code{\link[gamlss:gamlss]{gamlss}}.}
}
\value{
Returns a gamlss object. See the \strong{Value} section of
\code{\link[gamlss:gamlss]{gamlss}}. The class of the returned object is
\code{c("gamlssx", "gamlss", "gam", "glm", "lm")}.
}
\description{
Describe  \code{\link{GEV}}
}
\details{
Add details. Explain \code{stepAttempts} in more detail.
}
\examples{
# Simulate some data
set.seed(17012023)
n <- 100
x <- stats::runif(n)
mu <- 1 + 2 * x
sigma <- 1
xi <- 0.25
y <- nieve::rGEV(n = 1, loc = mu, scale = sigma, shape = xi)
plot(x, y)
data <- data.frame(y = as.numeric(y), x = x)
library(gamlss)

# Fit model using the default RS method
mod <- fitGEV(y ~ pb(x), data = data)

plot(mod)
plot(data$x, data$y)
lines(data$x, fitted(mod))
fitGEV(y ~ pb(x), data = data)

# Fit model using the mixed method
mod <- fitGEV(y ~ pb(x), data = data, method = mixed())

# Fit model using the CG method
mod <- fitGEV(y ~ pb(x), data = data, method = CG())
}
\seealso{
\code{\link{GEV}}, \code{\link[gamlss.dist:gamlss.family]{gamlss.family}},
\code{\link[gamlss:gamlss]{gamlss}}
}
