% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{GEV}
\alias{GEV}
\alias{dGEV}
\alias{pGEV}
\alias{qGEV}
\alias{rGEV}
\title{Generalized extreme value family distribution for fitting a GAMLSS}
\usage{
GEV(mu.link = "identity", sigma.link = "log", nu.link = "identity")

dGEV(x, mu = 0, sigma = 1, nu = 0, log = FALSE)

pGEV(q, mu = 0, sigma = 1, nu = 0, lower.tail = TRUE, log.p = FALSE)

qGEV(p, mu = 0, sigma = 1, nu = 0, lower.tail = TRUE, log.p = FALSE)

rGEV(n, mu = 0, sigma = 1, nu = 0)
}
\arguments{
\item{mu.link}{Defines the \code{mu.link}, with \code{"identity"} link as the default
for the \code{mu} parameter.}

\item{sigma.link}{Defines the \code{sigma.link}, with \code{"log"} link as the default
for the \code{sigma} parameter.}

\item{nu.link}{Defines the \code{nu.link}, with \code{"identity"} link as the default
for the \code{nu} parameter.}

\item{x, q}{Vector of quantiles.}

\item{mu, sigma, nu}{Vectors of location, scale and shape parameter values.}

\item{log, log.p}{Logical. If \code{TRUE}, probabilities \verb{eqn\{p\}} are given as
\eqn{\log(p)}.}

\item{lower.tail}{Logical. If \code{TRUE} (the default), probabilities are
\eqn{P[X \leq x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken to
be the number required.}
}
\value{
\code{GEV()} returns a \code{\link[gamlss.dist:gamlss.family]{gamlss.family}}
object which can be used to fit a generalized extreme value distribution
in the \code{\link[gamlss:gamlss]{gamlss}} function. \code{dGEV()} gives the density,
\code{pGEV()} gives the distribution function, \code{qGEV()} gives the quantile
function, and \code{rGEV()} generates random deviates.
}
\description{
The function \code{GEV} defines the generalized extreme value family
distribution, a three parameter distribution, for a
\code{\link[gamlss.dist:gamlss.family]{gamlss.family}} object to be used in GAMLSS
fitting using the function \code{\link[gamlss:gamlss]{gamlss}}. The functions
\code{dGEV}, \code{pGEV}, \code{qGEV} and \code{rGEV} define the density, distribution function,
quantile function and random generation for the specific parameterization of
the generalized extreme value distribution given in details below.
}
\details{
Add details.
\itemize{
\item Refer to Chapter 3 of Coles (2001) and Jenkinson (1955).
\item GEV(\eqn{\mu, \sigma, \xi}). \eqn{\nu = \xi}.
\item Give the pdf and/or cdf.
\item Explain the initial estimates.
\item Add code for the mean and variance.
}
}
\examples{
# Simulate some data
set.seed(17012023)
n <- 100
x <- stats::runif(n)
mu <- 1 + 2 * x
sigma <- 1
xi <- 0.25
y <- nieve::rGEV(n = 1, loc = mu, scale = sigma, shape = xi)
plot(x, y)
# Fit model
data <- data.frame(y = as.numeric(y), x = x)
library(gamlss)
mod <- gamlss(y ~ pb(x), family = GEV, data = data)
plot(mod)
plot(data$x, data$y)
lines(data$x, fitted(mod))

# Converges
modRS <- gamlss(y ~ pb(x), family = GEV, data = data, method = RS())
# Throws an error: parameters out-of-bounds leads to Inf deviance
modCG <- gamlss(y ~ pb(x), family = GEV, data = data, method = CG())
modMixed <- gamlss(y ~ pb(x), family = GEV, data = data, method = mixed())
}
\seealso{
\code{\link[gamlss.dist:gamlss.family]{gamlss.family}},
\code{\link[gamlss:gamlss]{gamlss}}
}
